Use nitro-protocol message signing instead of ethers wallet
