+++ packages/nitro-client/src/utils/index.ts
@@ -1,5 +1,6 @@
 export * from './helpers';
 export * from './peer';
 export { Nitro } from './nitro';
+export * from './key-signer';
 export * from './types';
 export * from './test/constants';
