+++ packages/server/test-e2e/integration.test.ts
@@ -1,7 +1,7 @@
 import 'mocha';
 import assert from 'assert';
 import { expect } from 'chai';
-import { ethers, providers } from 'ethers';
+import { providers } from 'ethers';
 
 import {
   Client, MemStore, Metrics, P2PMessageService, utils, Destination, LedgerChannelInfo,
@@ -51,7 +51,7 @@ const INITIAL_VIRTUAL_CHANNEL_AMOUNT = 1_000_000;
 const ASSET = `0x${'00'.repeat(20)}`;
 
 async function createClient(actor: utils.Actor, contractAddresses: ContractAddresses): Promise<[Client, P2PMessageService, Metrics]> {
-  const signer = new ethers.Wallet(actor.privateKey);
+  const signer = new utils.KeySigner(actor.privateKey);
   const clientStore = await MemStore.newMemStore(signer);
 
   const clientPeerIdObj = await createPeerIdFromKey(hex2Bytes(actor.privateKey));
